allow for arbitrary isochrone metallicities grid
Before, in a few places in the `IsochroneBase` the actual
isochrones were hardcoded to use those included in the original
package. This then wouldn't work for any other provided isochrones
(which *is* supported) unless they were on the exact same grid.

This commit makes the isochrone reader check the directory first
and read out what files actually exist. This still assumes the
isochrone files have the same names (with their metallicity) as
before